Add Caesar Cipher test
